requirements.txt
numpy==1.24.4
PyAudio==0.2.14
pyrtlsdr==0.3.0
